+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Forecast.hs
@@ -1,0 +1,21 @@
+{-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Cardano.Ledger.Allegra.Forecast () where
+
+import Cardano.Ledger.Allegra (AllegraEra)
+import Cardano.Ledger.Shelley.API.Forecast (EraForecast (..), ShelleyEraForecast (..))
+import Cardano.Ledger.Shelley.Forecast
+
+instance EraForecast AllegraEra where
+  type Forecast t AllegraEra = ShelleyForecast t AllegraEra
+  mkForecast = mkShelleyForecast
+  poolDistrForecastL = sfPoolDistrL
+  maxBlockHeaderSizeForecastL = sfMaxBlockHeaderSizeL
+  maxBlockBodySizeForecastL = sfMaxBlockBodySizeL
+  protocolVersionForecastL = sfProtocolVersionL
+
+instance ShelleyEraForecast AllegraEra where
+  genDelegsForecastL = sfGenDelegsL
+  decentralizationForecastL = sfDecentralizationL
+  extraEntropyForecastL = sfExtraEntropyL