+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Forecast.hs
@@ -1,0 +1,21 @@
+{-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Cardano.Ledger.Alonzo.Forecast () where
+
+import Cardano.Ledger.Alonzo (AlonzoEra)
+import Cardano.Ledger.Shelley.API.Forecast (EraForecast (..), ShelleyEraForecast (..))
+import Cardano.Ledger.Shelley.Forecast
+
+instance EraForecast AlonzoEra where
+  type Forecast t AlonzoEra = ShelleyForecast t AlonzoEra
+  mkForecast = mkShelleyForecast
+  poolDistrForecastL = sfPoolDistrL
+  maxBlockHeaderSizeForecastL = sfMaxBlockHeaderSizeL
+  maxBlockBodySizeForecastL = sfMaxBlockBodySizeL
+  protocolVersionForecastL = sfProtocolVersionL
+
+instance ShelleyEraForecast AlonzoEra where
+  genDelegsForecastL = sfGenDelegsL
+  decentralizationForecastL = sfDecentralizationL
+  extraEntropyForecastL = sfExtraEntropyL